DF-10716 make file system 64 bit seek get_size get_pos support
Former-commit-id: f84f1c14a1cfd750728b82e01bb8e877cb94210e

+++ Sources/Internal/Compression/ZipCompressor.cpp
@@ -109,7 +109,7 @@ ZipFile::ZipFile(const FilePath& fileName)
         throw std::runtime_error("can't open archive file: " + fileName.GetStringValue());
     }
 
-    uint32 fileSize = zipData->file->GetSize();
+    uint64 fileSize = zipData->file->GetSize();
 
     zipData->archive.m_pIO_opaque = zipData->file.get();
     zipData->archive.m_pRead = &file_read_func;
