+++ Sources/Internal/Render/Image/LibWebPHelper.cpp
@@ -37,7 +37,7 @@ eErrorCode LibWebPHelper::ReadFile(const ScopedPtr<File>& infile, Vector<Image*>
     }
 
     infile->Seek(0, File::SEEK_FROM_START);
-    uint32 dataSize = infile->GetSize();
+    uint32 dataSize = static_cast<uint32>(infile->GetSize());
     uint8_t* data = new uint8_t[dataSize];
     SCOPE_EXIT
     {
@@ -174,7 +174,7 @@ DAVA::ImageInfo LibWebPHelper::GetImageInfo(const ScopedPtr<File>& infile) const
     DVASSERT(infile);
 
     infile->Seek(0, File::SEEK_FROM_START);
-    uint32 dataSize = infile->GetSize();
+    uint32 dataSize = static_cast<uint32>(infile->GetSize());
     uint8_t* data = new uint8_t[dataSize];
     SCOPE_EXIT
     {
