+++ Sources/Internal/Sound/FMODSoundSystem.cpp
@@ -170,7 +170,7 @@ SoundEvent* SoundSystem::CreateSoundEventByID(const FastName& eventName, const F
 SoundEvent* SoundSystem::CreateSoundEventFromFile(const FilePath& fileName, const FastName& groupName, uint32 flags /* = SOUND_EVENT_DEFAULT */, int32 priority /* = 128 */)
 {
     SoundEvent* event = nullptr;
-    
+
 #ifdef __DAVAENGINE_IPHONE__
     if ((flags & SoundEvent::SOUND_EVENT_CREATE_STREAM) && !(flags & SoundEvent::SOUND_EVENT_CREATE_3D))
     {
@@ -709,7 +709,7 @@ void SoundSystem::RemoveSoundEventFromGroups(SoundEvent* event)
 
     soundGroupsMutex.Unlock();
 }
-    
+
 #ifdef __DAVAENGINE_IPHONE__
 bool SoundSystem::IsSystemMusicPlaying()
 {
@@ -731,7 +731,7 @@ FMOD_RESULT F_CALLBACK DAVA_FMOD_FILE_OPENCALLBACK(const char* name, int unicode
     if (!file)
         return FMOD_ERR_FILE_NOTFOUND;
 
-    (*filesize) = file->GetSize();
+    (*filesize) = static_cast<uint32>(file->GetSize());
     (*handle) = file;
 
     return FMOD_OK;
